# resource/PowerCLILaunch_db/Result.js
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE Result.js PLEASE EDIT ../custom/ResultCustom.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
const app = require('../../app.js');
const db_PowerCLILaunch_db = require('../../db/PowerCLILaunch_db_schema.js')
const properties = require('../../properties.js');
const handleError = require('../../security/util.js').handleError;
require('./custom/ResultCustom.js');

/*
 * SCHEMA DB Result
 * 
	{
		ExecTime: {
			type: 'Date'
		},
		Name: {
			type: 'String'
		},
		Value: {
			type: 'String'
		},
		//RELATIONS
		
		
		//EXTERNAL RELATIONS
		
		ScriptExecution: {
			type: Schema.ObjectId,
			ref : "Result"
		},
		
	}
 * 
 */



//CRUD METHODS


/**
 * ResultService.create
 *   @description CRUD ACTION create
 *
 */
app.post(properties.api + '/results', function (req, res) {
	obj = new db_PowerCLILaunch_db.Result(req.body);
	obj.save(function (err) {
		if (err) return handleError(err, res);
		res.send(obj);
	});
});

/**
 * ResultService.delete
 *   @description CRUD ACTION delete
 *   @param ObjectId id Id
 *
 */
app['delete'](properties.api + '/results/:id', function (req, res) {
	db_PowerCLILaunch_db.Result.findByIdAndRemove(req.params.id, function (err) {
		  if (err) return handleError(err, res);
		  res.send(err);
	});
});

/**
 * ResultService.findByScriptExecution
 *   @description CRUD ACTION findByScriptExecution
 *   @param String key
 *
 */
app.get(properties.api + '/results/findByScriptExecution/:key', function (req, res) {

	db_PowerCLILaunch_db.Result.find({ 'ScriptExecution' : req.params.key}).exec(function (err, list) {
		if (err) return handleError(err, res);
		res.send(list);
	});
	
});

/**
 * ResultService.update
 *   @description CRUD ACTION update
 *   @param ObjectId id Id
 *
 */
app.post(properties.api + '/results/:id', function (req, res) {
	db_PowerCLILaunch_db.Result.findByIdAndUpdate(req.params.id, req.body, {'new': true}, function (err, obj) {
		if (err) return handleError(err, res);
		res.send(obj);
	});
});


/*
 * CUSTOM SERVICES
 * 
 *	These services will be overwritten and implemented in  Custom.js
 */