# resource/PowerCLILaunch_db/ScriptExecution.js
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE ScriptExecution.js PLEASE EDIT ../custom/ScriptExecutionCustom.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
const app = require('../../app.js');
const db_PowerCLILaunch_db = require('../../db/PowerCLILaunch_db_schema.js')
const properties = require('../../properties.js');
const handleError = require('../../security/util.js').handleError;
require('./custom/ScriptExecutionCustom.js');

/*
 * SCHEMA DB ScriptExecution
 * 
	{
		Filename: {
			type: 'String', 
			required : true
		},
		HasResults: {
			type: 'String'
		},
		ResultType: {
			type: 'String'
		},
		State: {
			type: 'String'
		},
		TimeEnd: {
			type: 'Date'
		},
		TimeStart: {
			type: 'Date'
		},
		TimeUpdate: {
			type: 'Date'
		},
		//RELATIONS
		
		
		//EXTERNAL RELATIONS
		
		ScriptExecution: {
			type: Schema.ObjectId,
			ref : "Result"
		},
		
	}
 * 
 */



//CRUD METHODS


/**
 * ScriptExecutionService.create
 *   @description CRUD ACTION create
 *
 */
app.post(properties.api + '/scriptexecutions', function (req, res) {
	obj = new db_PowerCLILaunch_db.ScriptExecution(req.body);
	obj.save(function (err) {
		if (err) return handleError(err, res);
		res.send(obj);
	});
});

/**
 * ScriptExecutionService.delete
 *   @description CRUD ACTION delete
 *   @param ObjectId id Id
 *
 */
app['delete'](properties.api + '/scriptexecutions/:id', function (req, res) {
	db_PowerCLILaunch_db.ScriptExecution.findByIdAndRemove(req.params.id, function (err) {
		  if (err) return handleError(err, res);
		  res.send(err);
	});
});

/**
 * ScriptExecutionService.get
 *   @description CRUD ACTION get
 *   @param ObjectId id Id 
 *
 */
app.get(properties.api + '/scriptexecutions/:id', function (req, res) {
	db_PowerCLILaunch_db.ScriptExecution.findOne({_id:req.params.id}).exec(function (err, obj) {
		if (err) return handleError(err, res);
		res.send(obj);
	});
});

/**
 * ScriptExecutionService.list
 *   @description CRUD ACTION list
 *
 */
app.get(properties.api + '/scriptexecutions', function (req, res) {
	db_PowerCLILaunch_db.ScriptExecution.find().exec(function (err, list) {
		if (err) return handleError(err, res);
		res.send(list);
	});
});

/**
 * ScriptExecutionService.update
 *   @description CRUD ACTION update
 *   @param ObjectId id Id
 *
 */
app.post(properties.api + '/scriptexecutions/:id', function (req, res) {
	db_PowerCLILaunch_db.ScriptExecution.findByIdAndUpdate(req.params.id, req.body, {'new': true}, function (err, obj) {
		if (err) return handleError(err, res);
		res.send(obj);
	});
});


/*
 * CUSTOM SERVICES
 * 
 *	These services will be overwritten and implemented in  Custom.js
 */


/**
 * ScriptExecutionService.Launch
 *   @param String id
 *   @returns Boolean
 *
 */
app['get'](properties.api + '/scriptexecutions/launch/:id', function (req, res) {
	res.send({});
});
		